remove separate user step and make session own step
