+++ app/Filament/Pages/Installer/Steps/EnvironmentStep.php
@@ -3,20 +3,13 @@
 namespace App\Filament\Pages\Installer\Steps;
 
 use App\Filament\Pages\Installer\PanelInstaller;
+use Filament\Forms\Components\Fieldset;
 use Filament\Forms\Components\TextInput;
-use Filament\Forms\Components\ToggleButtons;
 use Filament\Forms\Components\Wizard\Step;
 use Filament\Forms\Set;
 
 class EnvironmentStep
 {
-    public const SESSION_DRIVERS = [
-        'file' => 'Filesystem',
-        'database' => 'Database',
-        'cookie' => 'Cookie',
-        'redis' => 'Redis',
-    ];
-
     public static function make(PanelInstaller $installer): Step
     {
         return Step::make('environment')
@@ -36,18 +29,26 @@ class EnvironmentStep
                     ->required()
                     ->default(url(''))
                     ->live()
-                    ->afterStateUpdated(fn ($state, Set $set) => $set('env_general.SESSION_SECURE_COOKIE', str_starts_with($state, 'https://') ? 'true' : 'false')),
-                TextInput::make('env_general.SESSION_SECURE_COOKIE')
-                    ->hidden()
-                    ->default(str_starts_with(url(''), 'https://') ? 'true' : 'false'),
-                ToggleButtons::make('env_session.SESSION_DRIVER')
-                    ->label('Session Driver')
-                    ->hintIcon('tabler-question-mark')
-                    ->hintIconTooltip('The driver used for storing sessions. We recommend "Filesystem" or "Database".')
-                    ->required()
-                    ->inline()
-                    ->options(self::SESSION_DRIVERS)
-                    ->default(config('session.driver')),
+                    ->afterStateUpdated(fn ($state, Set $set) => $set('env_session.SESSION_SECURE_COOKIE', str_starts_with($state, 'https://') ? 'true' : 'false')),
+                Fieldset::make('adminuser')
+                    ->label('Admin User')
+                    ->columns(3)
+                    ->schema([
+                        TextInput::make('user.email')
+                            ->label('E-Mail')
+                            ->required()
+                            ->email()
+                            ->placeholder('admin@example.com'),
+                        TextInput::make('user.username')
+                            ->label('Username')
+                            ->required()
+                            ->placeholder('admin'),
+                        TextInput::make('user.password')
+                            ->label('Password')
+                            ->required()
+                            ->password()
+                            ->revealable(),
+                    ]),
             ])
             ->afterValidation(fn () => $installer->writeToEnv('env_general'));
     }
